+++ changelog.md
@@ -4,12 +4,14 @@
 
 ### Version 5.0 - 2024.07.08
 
+* New: updated library and plugins system requirements
 * New: extracted `Dialogs` JavaScript library from `Admin`
 * New: added several more icons to the icon font
 * Edit: default location is now `library`
 * Edit: many updates to the `Background` class
 * Edit: many updates to the `CRON` task class
 * Removed: outdated JavaScript `helpers` library
+* Removed: all previously deprecated methods
 
 ### Version 4.9.2 - 2024.06.26